requirements.txt
aiogram==2.11.2
aiohttp==3.7.4
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
cached-property==1.5.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
gino==1.0.1
hiredis==1.1.0
idna==2.10
instaloader==4.6.2
mock==4.0.3
multidict==5.1.0
peewee==2.10.2
peewee-async==0.5.12
peewee-migrate==0.14.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
